Zona de administrador

+++ src/Controller/ControladorUsuario.php
@@ -20,9 +20,7 @@ class ControladorUsuario extends AbstractController
     public function perfilUsuario(Request $request, EntityManagerInterface $entityManager)
     {
         if ($this->getUser() == null) {
-            return $this->render("anuncio.html.twig", [
-                "mensaje" => "Si no estás logeado, no puedes acceder :("
-            ]);
+            return $this->render("anuncio_error.html.twig", [ "mensaje" => "Si no estás logeado, no puedes acceder :(" ]);
         }
         else {
             //Obtener el usuario actual
@@ -47,57 +45,54 @@ class ControladorUsuario extends AbstractController
     #[Route('/procesar_editar_usuario', name: 'procesar_editar_usuario')]
     public function procesar_editar_usuario(Request $request, EntityManagerInterface $entityManager, UserPasswordHasherInterface $passwordHasher)
     {
-        // Recoger datos del formulario tras la verificación del usuario
-        $nombre = $request->request->get('nombre');
-        $apellidos = $request->request->get('apellidos');
-        $correo = $request->request->get('email');
-        $direccion = $request->request->get('direccion');
-        $contrasena = $request->request->get('contrasena');
-        $confirmar_contrasena = $request->request->get('confirmar_contrasena');
-
-        //Recoger el usuario que queremos modificar (el propio que está logeado, claramente)
-        $usuario_identifier = $this->getUser()->getUserIdentifier();
-        $usuario = $entityManager->getRepository(Usuario::class)->findOneBy(['email' => $usuario_identifier]);
-        
         if ($this->getUser() == null) {
-            return $this->render("anuncio.html.twig", [
-                "titulo_mensaje" => "Vaya...",
-                "mensaje" => "Lo sentimos... si no estás logeado, no puedes acceder :("
-            ]);
+            return $this->render("anuncio_error.html.twig", [ "mensaje" => "Si no estás logeado, no puedes acceder :(" ]);
         }
         else {
-            // Aquí empieza la fase para verificar qué campos están vacíos y cuáles no, en caso de estarlo, se entiende que el usuario no quiere modificar nada.
-            if ($nombre != "") {
-                $usuario->setNombreUsuario($nombre);
+            // Recoger datos del formulario tras la verificación del usuario
+            $nombre = $request->request->get('nombre');
+            $apellidos = $request->request->get('apellidos');
+            $correo = $request->request->get('email');
+            $direccion = $request->request->get('direccion');
+            $contrasena = $request->request->get('contrasena');
+            $confirmar_contrasena = $request->request->get('confirmar_contrasena');
+
+            //Recoger el usuario que queremos modificar (el propio que está logeado, claramente)
+            $usuario_identifier = $this->getUser()->getUserIdentifier();
+            $usuario = $entityManager->getRepository(Usuario::class)->findOneBy(['email' => $usuario_identifier]);
+            
+            if ($this->getUser() == null) {
+                return $this->render("anuncio_error.html.twig", [ "mensaje" => "Lo sentimos... si no estás logeado, no puedes acceder :(" ]);
             }
-            if ($apellidos != "") {
-                $usuario->setApellidoUsuario($apellidos);
-            }
-            if ($correo != "") {
-                $usuario_repetido = $entityManager->getRepository(Usuario::class)->findOneBy(['email' => $correo]);
-                if ($usuario_repetido) {
-                    return $this->render("anuncio.html.twig", [
-                        "titulo_mensaje" => "Vaya...",
-                        "mensaje" => "Lo sentimos... pero este email ya está escogido"
-                    ]);
+            else {
+                // Aquí empieza la fase para verificar qué campos están vacíos y cuáles no, en caso de estarlo, se entiende que el usuario no quiere modificar nada.
+                if ($nombre != "") {
+                    $usuario->setNombreUsuario($nombre);
                 }
-                $usuario->setEmail($correo);
-            }
-            if ($direccion != "") {
-                $usuario->setDireccion($direccion);
-            }
-            if ($contrasena != "") {
-                $hashedPassword = $passwordHasher->hashPassword($usuario, $contrasena);
-                $usuario->setPassword($hashedPassword);
-            }
+                if ($apellidos != "") {
+                    $usuario->setApellidoUsuario($apellidos);
+                }
+                if ($correo != "") {
+                    $usuario_repetido = $entityManager->getRepository(Usuario::class)->findOneBy(['email' => $correo]);
+                    if ($usuario_repetido) {
+                        return $this->render("anuncio_error.html.twig", [ "mensaje" => "Lo sentimos... pero este email ya está escogido" ]);
+                    }
+                    $usuario->setEmail($correo);
+                }
+                if ($direccion != "") {
+                    $usuario->setDireccion($direccion);
+                }
+                if ($contrasena != "") {
+                    $hashedPassword = $passwordHasher->hashPassword($usuario, $contrasena);
+                    $usuario->setPassword($hashedPassword);
+                }
 
-            $entityManager->persist($usuario);
-            $entityManager->flush();
+                $entityManager->persist($usuario);
+                $entityManager->flush();
 
-            return $this->render("anuncio.html.twig", [
-                "titulo_mensaje" => "¡Listo!",
-                "mensaje" => "Hemos terminado de editar tus datos, puedes volver a tu usuario o de vuelta a la tienda a través de la barra superior ;D"
-            ]);
+                return $this->render("anuncio_exito.html.twig", [ "mensaje" => "Hemos terminado de editar tus datos, puedes volver a tu usuario o de vuelta a la tienda a través de la barra superior ;D" ]);
+        
+            }
         }
     }
 
@@ -108,20 +103,21 @@ class ControladorUsuario extends AbstractController
         $orden = $entityManager->getRepository(Orden::class)->findOneBy(['id_orden' => $id_orden]);
 
         if ($this->getUser() == null) {
-            return $this->render("anuncio.html.twig", [
-                "mensaje" => "Si no estás logeado, no puedes acceder :("
-            ]);
+            return $this->render("anuncio_error.html.twig", [ "mensaje" => "Si no estás logeado, no puedes acceder :(" ]);
         }
         else {
+
+            if ($request->isMethod('GET')) {
+                return $this->render("anuncio_error.html.twig", [
+                    "mensaje" => "Debes usar el formulario para cancelar un pedido."
+                ]);
+            }
             
             $orden->setEstado("Cancelado");
             $entityManager->persist($orden);
             $entityManager->flush();
 
-            return $this->render("anuncio.html.twig", [
-                "titulo_mensaje" => "¡Listo!",
-                "mensaje" => "Hemos cancelado el pedido con éxito ;D"
-            ]);
+            return $this->render("anuncio_exito.html.twig", [ "mensaje" => "Hemos cancelado el pedido con éxito ;D" ]);
         }
     }
 
@@ -132,51 +128,56 @@ class ControladorUsuario extends AbstractController
         $tarjeta = $entityManager->getRepository(Tarjeta::class)->findOneBy(['idTarjeta' => $id_tarjeta]);
 
         if ($this->getUser() == null) {
-            return $this->render("anuncio.html.twig", [
-                "mensaje" => "Si no estás logeado, no puedes acceder :("
-            ]);
+            return $this->render("anuncio_error.html.twig", [ "mensaje" => "Si no estás logeado, no puedes acceder :(" ]);
         }
         else {
+            if ($request->isMethod('GET')) {
+                return $this->render("anuncio_error.html.twig", [
+                    "mensaje" => "Debes usar el formulario para eliminar una tarjeta."
+                ]);
+            }
             
             $entityManager->remove($tarjeta);
             $entityManager->flush();
 
-            return $this->render("anuncio.html.twig", [
-                "titulo_mensaje" => "¡Listo!",
-                "mensaje" => "Hemos eliminado esta tarjeta ;D"
-            ]);
+            return $this->render("anuncio_exito.html.twig", [ "mensaje" => "Hemos eliminado esta tarjeta ;D" ]);
         }
     }
 
     #[Route('/procesar_anadir_tarjeta', name: 'procesar_anadir_tarjeta')]
     public function procesar_anadir_tarjeta(Request $request, EntityManagerInterface $entityManager)
     {   
-        $usuario_identifier = $this->getUser()->getUserIdentifier();
-        $id_usuario = $entityManager->getRepository(Usuario::class)->findOneBy(['email' => $usuario_identifier]);
-        $titular_tarjeta = $request->request->get('titular');
-        $numero_tarjeta = $request->request->get('numero_tarjeta');
-        $ccv = $request->request->get('ccv');
-        $fecha_expiracion = $request->request->get('fecha_expiracion');
-
         if ($this->getUser() == null) {
-            return $this->render("anuncio.html.twig", [
-                "mensaje" => "Si no estás logeado, no puedes acceder :("
-            ]);
+            return $this->render("anuncio_error.html.twig", [ "mensaje" => "Si no estás logeado, no puedes acceder :(" ]);
         }
         else {
-            $tarjeta = new Tarjeta();
-            $tarjeta->setNombreTitular($titular_tarjeta);
-            $tarjeta->setUsuario($id_usuario);
-            $tarjeta->setNumeroTarjeta($numero_tarjeta);
-            $tarjeta->setCcv($ccv);
-            $tarjeta->setFechaExpiracion($fecha_expiracion);
-            $entityManager->persist($tarjeta);
-            $entityManager->flush();
+            if ($request->isMethod('GET')) {
+                return $this->render("anuncio_error.html.twig", [
+                    "mensaje" => "Debes usar el formulario para añadir una tarjeta."
+                ]);
+            }
+            $usuario_identifier = $this->getUser()->getUserIdentifier();
+            $id_usuario = $entityManager->getRepository(Usuario::class)->findOneBy(['email' => $usuario_identifier]);
+            $titular_tarjeta = $request->request->get('titular');
+            $numero_tarjeta = $request->request->get('numero_tarjeta');
+            $ccv = $request->request->get('ccv');
+            $fecha_expiracion = $request->request->get('fecha_expiracion');
 
-            return $this->render("anuncio.html.twig", [
-                "titulo_mensaje" => "¡Listo!",
-                "mensaje" => "Hemos añadido esta tarjeta a tu cuenta ;D"
-            ]);
+            if ($this->getUser() == null) {
+                return $this->render("anuncio_error.html.twig", [ "mensaje" => "Si no estás logeado, no puedes acceder :(" ]);
+            }
+            else {
+                $tarjeta = new Tarjeta();
+                $tarjeta->setNombreTitular($titular_tarjeta);
+                $tarjeta->setUsuario($id_usuario);
+                $tarjeta->setNumeroTarjeta($numero_tarjeta);
+                $tarjeta->setCcv($ccv);
+                $tarjeta->setFechaExpiracion($fecha_expiracion);
+                $entityManager->persist($tarjeta);
+                $entityManager->flush();
+
+                return $this->render("anuncio_exito.html.twig", [ "mensaje" => "Hemos añadido esta tarjeta a tu cuenta ;D" ]);
+            }
         }
   
     }
